Set required php to 8.2 or above and add trait for duplicate constants and functions.
