+++ composer.json
@@ -16,7 +16,7 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": "^8.1",
+    "php": "^8.2",
     "laravel/framework": "^10.0",
     "laravel/slack-notification-channel": "^2.0",
     "eveseat/eveapi": "^5.0",
